Initial form of a "Session Store"
Right now the "session" is a hardcoded dict in the session storage type
module, but it could easily be adapted to be a JSON file.
Highly handicapped at present, built only for `transfer.endpoint_id`.
Converts `ls` to have this style of config-loaded required option. We
may want to customize helptext on these (future todo).
Also converts PATH to be a variadic argument so you can `globus transfer
ls` or `globus transfer ls /abc/`