Convert "SessionLoadedOption" to Config Loaded
This is more than a naming change. The idea that these will be loaded
from a separate "Session" file is gone. Instead, we'll use a few config
sections in ~/.globus.cfg to track these variables as persistent values.
Updates `globus config` commands to set/get values in the [cli],
[cli transfer], and any future [cli] subsections. This lets us use the
comfortable dot-based notation to describe values as they're written
into the config.
Also convert `--format` to be a ConfigLoadedOption, since that's more
uniform. Still uses CommandState to track the actual resolved-value
post-callback, so that we can persist the `expose_value=False` and
associated "clean" semantics.
Open question: should all ConfigLoadedOptions be stored on the
CommandState in some organized manner, and fetched therewith much as the
output format is? Seems like it would reduce the bloat attached to
functions consuming the endpoint ID, in that they no longer get a
magical argument passed in.