Add command aliases as a type of click.Command
The aliases are constructed in terms of a rename and an existing command
object, and assign nearly all of the same attributes to the alias
command. The only exception is the name, which is obviously altered in
the process.
Importantly, this is a type of click.Command, and is not generic enough
to safely apply to a Group or MultiCommand -- the attributes for those
types are different, and the passthrough of original command attributes
to the new command's constructor will not cover the differences.

The AliasCommand importantly inherits the original commands params, so
it does not need to blindly forward options as unprocessed values. That
saves us from some of the concerns over doing that.

Make
 - `globus endpoint-search` -> `globus transfer endpoint search`
 - `globus endpoint-ls` -> `globus transfer ls`
as the first two such aliases.

Resolves #71