Add experimental "completion cache" file
This adds a somewhat questionable kind of completion, for now just doing
bookmark names on `bookmark locate`. The basic idea is that for simple
lists like this, the values can be fetched periodically and cached on
the local filesystem. The completer can then lookup those values and use
them as completion options, rewriting the cache file if it's ever stale.
Dealing with bookmark names forced me to address some of the issues
regarding escaping/quoting already present in the completion support.

The other interesting pattern here is the idea of a command having a
custom completion func attached to it, which is invoked when none of the
option matching or other rules apply.