Add initial form of `--man`, use it on `transfer`
To get the `--man` flag available on a command, we need to actually make
it a specialized subclass of click.Command which tracks some additional
manpage data. Several sections of the man output can be built purely
from helptext and basic command name data, but some require custom text
(like DESCRIPTION and EXAMPLE). For this purpose, just put Groff
formatted text directly into the command definition. It's a bit messy
because we suddenly have ugly groff nonsense directly in the command,
but the only alternative is to build more things which _render to_
groff. That seems more complex and fragile, and therefore worse.