Add IdentityType ParamType used in get-identities
This is a new parameter type which encodes what it means to parse an
incoming identity or domain. It's dropped into place for
`get-identities` and it will validate and parse those inputs as
before. Unlike the prior validation, it will reject invalid identity
names which do not contain '@', since these cannot possibly be
usernames.

The IdentityType supports the b32-encoded unix-compatible username
used by Transfer, so that detail is no longer encoded into the
get-identities module.

Because this is slightly stricter about validating that usernames
contain '@', it is possible that the command will fail with an
exit(2) in situations in which it would previously just produce
`NO_SUCH_IDENTITY`.