Add `description` to `endpoint show` text
And make the formatter a little bit more organized. Rather than doing
tuple-unpacking in the various printers, a new FormatField class is
added which can coerce a tuple to a FormatField (or passthrough any
existing FormatField). This lets us keep the existing tuple-based
interace but introduce fields with parameters. Namely, `description`
is unusual in that text-wrapping may be desirable, so it is defined as
a FormatField where `wrap_enabled=True`.

We then make a judgement while printing the data about "should wrap"
to allow for non-wrapping behavior for short outputs.

Description is truncated to a maximum of 5 lines of output, newlines
are respected, and the whole output is right-aligned to line up with
existing fields.

The table printer needed very minor tweaks to handle FormatField
instead of tuples.

When adding new tests for `endpoint show`, it turned out that our
fixture data (some originally taken from docs) is out of date and
doesn't have several supported fields. Simply pad out the missing
fields so that tests will be in good working order.