Allow FormatField to consume a str
When a string is given, rather than a tuple, it will be converted to a
title-cased version paired with the original. Underscores are
converted to spaces.

So `foo_bar` -> `(Foo Bar, foo_bar)`

This is then used for FormatField construction as before.

There isn't a huge net-reduction in lines of code, but it does reduce
the total character count a bit. Notably, the diff is almost entirely
negative, with only two logical lines added.