Allow fields to define themselves as conditional
A conditional field has some rendering criterion (`is_present`) which
can be evaluated by the formatters against the data being printed.
This allows fields to self-suppress when conditions are met. For
example, all of the filtering done on collection-conditional fields is
easily replaced with self-filtering fields.

This capability is only supported for record-formatted output. In
table output, there is no relevant record to check.